requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
mysqlclient==1.3.13
django-mysql==2.4.1
simplejson==3.16.0
django-markdown-app==0.9.4.1
qiniu==7.2.2
oss2==2.5.0
whitenoise==4.1
